feat: add types
