+++ src/routes/graphql/types/profile.ts
@@ -1,56 +1,18 @@
-import graphql from "graphql";
-import { UUIDType } from "../types/uuid.js";
 import { MemberType } from "./memberType.js";
-import { UserType } from "./user.js";
-import { PrismaClient } from "@prisma/client";
-import { MemberTypeId } from "./memberTypeId.js";
+import { User } from "./user.js";
 
-export let ProfileType: graphql.GraphQLObjectType;
-export function createProfileType(prisma: PrismaClient) {
-  ProfileType = new graphql.GraphQLObjectType({
-    name: 'Profile',
-    fields: () => ({
-      id: {type: new graphql.GraphQLNonNull(UUIDType)},
-      isMale: {type: new graphql.GraphQLNonNull(graphql.GraphQLBoolean)},
-      yearOfBirth: {type: new graphql.GraphQLNonNull(graphql.GraphQLInt)},
-      userId: {type: new graphql.GraphQLNonNull(UUIDType)},
-      user: {
-        type: UserType,
-        resolve: (parent) => {
-          return prisma.user.findFirst({
-            where: {
-              id: parent.userId
-            }
-          })
-        }
-      },
-      memberTypeId: {type: new graphql.GraphQLNonNull(MemberTypeId)},
-      memberType: {
-        type: MemberType,
-        resolve: async (source, args, context) => {
-          const memberType = await context.loaders.memberTypeLoader.load(source.memberTypeId)
-          return memberType 
-        }
-      }
-    }),
-  })
+export type ProfileInput = {
+  id: string;
+  dto: any;
+  userId: string;
 }
 
-
-export const CreateProfileInput = new graphql.GraphQLInputObjectType({
-  name: 'CreateProfileInput',
-  fields: {
-    userId: {type: new graphql.GraphQLNonNull(UUIDType)},
-    memberTypeId: {type: new graphql.GraphQLNonNull(MemberTypeId)},
-    isMale: {type: graphql.GraphQLBoolean},
-    yearOfBirth: {type: graphql.GraphQLInt},
-  },
-})
-
-export const ChangeProfileInput = new graphql.GraphQLInputObjectType({
-  name: 'ChangeProfileInput',
-  fields:() => ({
-    isMale: {type: graphql.GraphQLBoolean},
-    yearOfBirth: {type: graphql.GraphQLInt},
-  }),
-})
+export type Profile = {
+  id: string;
+  isMale: boolean;
+  yearOfBirth: number;
+  userId: string;
+  user: User;
+  memberTypeId: string;
+  memberType: MemberType;
+}
